split structs into sample and paramter strcuts
