+++ Diagnostics/CombineMaCh3Chains.cpp
@@ -3,7 +3,7 @@
 
 // MaCh3 includes
 #include "Manager/Manager.h"
-#include "Samples/Structs.h"
+#include "Samples/SampleStructs.h"
 
 _MaCh3_Safe_Include_Start_ //{
 // ROOT includes
