+++ Diagnostics/GetPenaltyTerm.cpp
@@ -1,6 +1,6 @@
 // MaCh3 includes
 #include "Manager/Manager.h"
-#include "Samples/Structs.h"
+#include "Samples/SampleStructs.h"
 
 _MaCh3_Safe_Include_Start_ //{
 // ROOT includes
