+++ Fitters/StatisticalUtils.h
@@ -9,23 +9,16 @@
 #include <cmath>
 
 //MaCh3 includes
-#include "Samples/Structs.h"
 #include "Manager/Manager.h"
 #include "Samples/HistogramUtils.h"
 
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wuseless-cast"
-#pragma GCC diagnostic ignored "-Wfloat-conversion"
-#pragma GCC diagnostic ignored "-Wfloat-conversion"
-#pragma GCC diagnostic ignored "-Wold-style-cast"
-#pragma GCC diagnostic ignored "-Wconversion"
-#pragma GCC diagnostic ignored "-Wformat-nonliteral"
+_MaCh3_Safe_Include_Start_ //{
 // ROOT includes
 #include "TCanvas.h"
 #include "TLine.h"
 #include "TROOT.h"
 #include "TStyle.h"
-#pragma GCC diagnostic pop
+_MaCh3_Safe_Include_End_ //}
 
 /// @file StatisticalUtils.h
 /// @brief Utility functions for statistical interpretations in MaCh3
