+++ Parameters/ParameterHandlerGeneric.h
@@ -2,7 +2,7 @@
 
 // MaCh3 includes
 #include "Parameters/ParameterHandlerBase.h"
-#include "Samples/Structs.h"
+#include "Samples/SampleStructs.h"
 
 /// @brief Class responsible for handling of systematic error parameters with different types defined in the config. Like spline, normalisation parameters etc.
 /// @see For more details, visit the [Wiki](https://github.com/mach3-software/MaCh3/wiki/02.-Implementation-of-Systematic).
