+++ Plotting/GetPostfitParamPlots.cpp
@@ -5,7 +5,6 @@
 
 #include "plottingUtils/plottingUtils.h"
 #include "plottingUtils/plottingManager.h"
-#include "Samples/Structs.h"
 
 _MaCh3_Safe_Include_Start_ //{
 #include "TROOT.h"
