+++ Samples/HistogramUtils.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // MaCh3 includes
-#include "Samples/Structs.h"
+#include "Samples/SampleStructs.h"
 
 _MaCh3_Safe_Include_Start_ //{
 // ROOT include
